api/requirements.txt
fastapi==0.112.0
uvicorn==0.23.2
tensorflow-cpu==2.20.0
numpy
pillow
